productmanagement/requirement.txt
asgiref==3.4.1
Django==3.2.25
djangorestframework==3.14.0
pkg-resources==0.0.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.1.1
